requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
redis==5.0.3
python-dotenv==1.0.1